iota_types

Module base_types

Re-exports

pub use crate::committee::EpochId;

pub use crate::digests::ObjectDigest;

pub use crate::digests::TransactionDigest;

pub use crate::digests::TransactionEffectsDigest;

Structs

ExecutionData

ExecutionDigests

Identifier

A move identifier

IotaAddress

Unique identifier for an Account on the IOTA blockchain.

MoveLegacyTxContext

MoveObjectType

A StructTag with optimized BCS serialization for object types.

ObjectID

An ObjectId is a 32-byte identifier used to uniquely identify an object on the IOTA blockchain.

ObjectInfo

ObjectRef

Reference to an object

SequenceNumber

SizeOneVec

StructTag

Type information for a move struct

TxContext

VerifiedExecutionData

Enums

ObjectIDParseError

ObjectType

Type of an IOTA object

TxContextKind

TypeTag

Type of a move value

Constants

IOTA_ADDRESS_LENGTH

RESOLVED_ASCII_STR

RESOLVED_STD_OPTION

RESOLVED_UTF8_STR

Traits

ConciseableName

Functions

address_from_iota_pub_key

dbg_addr

Generate a fake IotaAddress with repeated one byte.

dbg_object_id

Generate a fake ObjectID with repeated one byte.

is_primitive_type_tag

Whether this type is valid as a primitive (pure) transaction input.

move_ascii_str_layout

move_utf8_str_layout

random_object_ref

url_layout

Type Aliases

AuthorityName

CommitRound

The round number.

TxSequenceNumber

VersionDigest

VersionNumber