iota_types::base_types

Struct StructTag

pub struct StructTag { /* private fields */ }

Type information for a move struct

BCS

The BCS serialized form for this type is defined by the following ABNF:

struct-tag = address            ; address of the package
             identifier         ; name of the module
             identifier         ; name of the type
             (vector type-tag)  ; type parameters

Implementations

impl StructTag

pub fn new( address: impl Into<Address>, module: impl Into<Identifier>, name: impl Into<Identifier>, type_params: Vec<TypeTag>, ) -> StructTag

pub fn is_same_type_as(&self, other: &StructTag) -> bool

Returns whether the type (excluding the type params) is the same as another struct tag.

pub fn new_gas() -> StructTag

Creates a new framework IOTA gas struct tag (0x2::iota::IOTA)

pub fn is_gas(&self) -> bool

Checks if this is a framework IOTA gas type (0x2::iota::IOTA)

pub fn new_gas_coin() -> StructTag

Creates a new framework gas coin struct tag (0x2::coin::Coin<0x2::iota::IOTA>)

pub fn is_gas_coin(&self) -> bool

Checks if this is a framework gas coin type (0x2::coin::Coin<0x2::iota::IOTA>)

pub fn new_id() -> StructTag

Creates a new framework object ID struct tag (0x2::object::ID)

pub fn is_id(&self) -> bool

Checks if this is a framework object ID type (0x2::object::ID)

pub fn new_uid() -> StructTag

Creates a new framework object UID struct tag (0x2::object::UID)

pub fn is_uid(&self) -> bool

Checks if this is a framework object UID type (0x2::object::UID)

pub fn new_name(address: Address) -> StructTag

Creates a new IOTA-Names Name struct tag with the given package address.

let iota_names_config = IotaNamesConfig::mainnet();
let package_address = iota_names_config.package_address;
assert_eq!(
    StructTag::new_name(package_address).to_canonical_string(true),
    format!("{package_address}::name::Name")
);

pub fn is_name(&self) -> bool

Checks if this is an IOTA-Names Name type. Note that this does not check the package address, so it will return true for any struct with the correct module and type name with no type params.

pub fn new_dynamic_field( key: impl Into<TypeTag>, value: impl Into<TypeTag>, ) -> StructTag

Creates a new dynamic field struct tag (0x2::dynamic_field::Field<KeyType, ValueType>)

pub fn is_dynamic_field(&self) -> bool

Checks if this is a Dynamic Field type (0x2::dynamic_field::Field<KeyType, ValueType>)

pub fn coin_type_opt(&self) -> Option<&TypeTag>

Returns the coin type if this is a Coin type, None otherwise

pub fn coin_type(&self) -> &TypeTag

Returns the coin type of this StructTag, panics if not a coin

pub fn address(&self) -> Address

Returns the address part of a StructTag

pub fn module(&self) -> &Identifier

Returns the module part of a StructTag

pub fn name(&self) -> &Identifier

Returns the name part of a StructTag

pub fn type_params(&self) -> &[TypeTag]

Returns the type params part of a StructTag

pub fn type_params_mut(&mut self) -> &mut Vec<TypeTag>

Returns a mutable reference to the type params part of a StructTag

pub fn into_parts(self) -> (Address, Identifier, Identifier, Vec<TypeTag>)

Decomposes the StructTag into its parts

pub fn is_timelocked_balance(&self) -> bool

Checks if this is a timelocked coin balance TimeLock<Balance<T>>

pub fn is_gas_balance(&self) -> bool

Checks if this is an IOTA balance type (0x2::balance::Balance<0x2::iota::IOTA>)

pub fn new_timelocked_gas_balance() -> StructTag

Creates a new timelocked IOTA balance struct tag (0x2::timelock::TimeLock<0x2::balance::Balance<0x2::iota::IOTA>>)

pub fn is_timelocked_gas_balance(&self) -> bool

Checks if this is a timelocked IOTA balance type (0x2::timelock::TimeLock<0x2::balance::Balance<0x2::iota::IOTA>>)

pub fn to_canonical_string(&self, with_prefix: bool) -> String

Returns the string representation of this struct tag using the canonical display, with or without a 0x prefix.

impl StructTag

pub fn new_authenticator_state() -> StructTag

pub fn is_authenticator_state(&self) -> bool

pub fn new_clock() -> StructTag

pub fn is_clock(&self) -> bool

pub fn new_config() -> StructTag

pub fn is_config(&self) -> bool

pub fn new_iota_treasury_cap() -> StructTag

pub fn is_iota_treasury_cap(&self) -> bool

pub fn new_upgrade_cap() -> StructTag

pub fn is_upgrade_cap(&self) -> bool

pub fn new_upgrade_ticket() -> StructTag

pub fn is_upgrade_ticket(&self) -> bool

pub fn new_upgrade_receipt() -> StructTag

pub fn is_upgrade_receipt(&self) -> bool

pub fn new_random() -> StructTag

pub fn is_random(&self) -> bool

pub fn new_iota_system_admin_cap() -> StructTag

pub fn is_iota_system_admin_cap(&self) -> bool

pub fn new_url() -> StructTag

pub fn is_url(&self) -> bool

pub fn new_bag() -> StructTag

pub fn is_bag(&self) -> bool

pub fn new_object_bag() -> StructTag

pub fn is_object_bag(&self) -> bool

pub fn new_tx_context() -> StructTag

pub fn is_tx_context(&self) -> bool

pub fn new_deny_list_config_key() -> StructTag

pub fn is_deny_list_config_key(&self) -> bool

pub fn new_deny_list_address_key() -> StructTag

pub fn is_deny_list_address_key(&self) -> bool

pub fn new_deny_list_global_pause_key() -> StructTag

pub fn is_deny_list_global_pause_key(&self) -> bool

pub fn new_transfer_receiving() -> StructTag

pub fn is_transfer_receiving(&self) -> bool

pub fn new_iota_system_state() -> StructTag

pub fn is_iota_system_state(&self) -> bool

pub fn new_staked_iota() -> StructTag

pub fn is_staked_iota(&self) -> bool

pub fn new_timelocked_staked_iota() -> StructTag

pub fn is_timelocked_staked_iota(&self) -> bool

pub fn new_system_epoch_info_event() -> StructTag

pub fn is_system_epoch_info_event(&self) -> bool

pub fn new_string() -> StructTag

pub fn is_string(&self) -> bool

pub fn new_ascii_string() -> StructTag

pub fn is_ascii_string(&self) -> bool

pub fn new_regulated_coin_metadata() -> StructTag

pub fn is_regulated_coin_metadata(&self) -> bool

pub fn new_deny_cap_v1() -> StructTag

pub fn is_deny_cap_v1(&self) -> bool

pub fn new_coin_metadata(struct_tag: impl Into<StructTag>) -> StructTag

pub fn is_coin_metadata(&self) -> bool

pub fn new_treasury_cap(struct_tag: impl Into<StructTag>) -> StructTag

pub fn is_treasury_cap(&self) -> bool

pub fn new_coin_manager(struct_tag: impl Into<StructTag>) -> StructTag

pub fn is_coin_manager(&self) -> bool

pub fn new_display_created(struct_tag: impl Into<StructTag>) -> StructTag

pub fn is_display_created(&self) -> bool

pub fn new_display_version_updated( struct_tag: impl Into<StructTag>, ) -> StructTag

pub fn is_display_version_updated(&self) -> bool

pub fn new_coin(type_tag: impl Into<TypeTag>) -> StructTag

pub fn is_coin(&self) -> bool

pub fn new_balance(type_tag: impl Into<TypeTag>) -> StructTag

pub fn is_balance(&self) -> bool

pub fn new_time_lock(type_tag: impl Into<TypeTag>) -> StructTag

pub fn is_time_lock(&self) -> bool

pub fn new_config_setting(type_tag: impl Into<TypeTag>) -> StructTag

pub fn is_config_setting(&self) -> bool

pub fn new_dynamic_object_field_wrapper( type_tag: impl Into<TypeTag>, ) -> StructTag

pub fn is_dynamic_object_field_wrapper(&self) -> bool

pub fn new_option(type_tag: impl Into<TypeTag>) -> StructTag

pub fn is_option(&self) -> bool