Module iota_types::committee

Structs

Committee

CommitteeWithNetworkMetadata

NetworkMetadata

ProtocolVersion

Constants

QUORUM_THRESHOLD

Quorum threshold for our fixed voting power–any message signed by this much voting power can be trusted up to BFT assumptions

TOTAL_VOTING_POWER

Set total_voting_power as 10_000 by convention. Individual voting powers can be interpreted as easily understandable basis points (e.g., voting_power: 100 = 1%, voting_power: 1 = 0.01%). Fixing the total voting power allows clients to hardcode the quorum threshold and total_voting power rather than recomputing these.

VALIDITY_THRESHOLD

Validity threshold defined by f+1

Traits

CommitteeTrait

Type Aliases

CommitteeDigest

EpochId

StakeUnit