iota_types

Module effects

Structs

EffectsObjectChange

Input/output state of an object that was changed during execution

ObjectChange

Effect on an individual object, keyed by its [ObjectId].

TestEffectsBuilder

TransactionEffectsDebugSummary

TransactionEffectsV1

Version 1 of TransactionEffects

TransactionEvents

Events emitted during the successful execution of a transaction

Enums

IDOperation

Defines what happened to an ObjectId during execution

InputSharedObject

Description of a shared object that was used as input to a transaction.

ObjectIn

State of an object prior to execution

ObjectOut

State of an object after execution

ObjectRemoveKind

TransactionEffects

The output or effects of executing a transaction

Constants

APPROX_SIZE_OF_EPOCH_ID

APPROX_SIZE_OF_EXECUTION_STATUS

APPROX_SIZE_OF_GAS_COST_SUMMARY

APPROX_SIZE_OF_OBJECT_REF

APPROX_SIZE_OF_OPT_TX_EVENTS_DIGEST

APPROX_SIZE_OF_OWNER

APPROX_SIZE_OF_TX_DIGEST

Traits

TransactionEffectsAPI

Version-agnostic accessors for TransactionEffects.

TransactionEffectsAPIForTesting

Test-only mutators for TransactionEffects that bypass the normal invariants. Not for production use.

TransactionEffectsExt

The version-selecting constructor and aggregating queries for the TransactionEffects enum. Sealed; implemented only for the enum.

TransactionEffectsExtForTesting

Test-only counterpart to TransactionEffectsExt. Not for production use.

Type Aliases

CertifiedTransactionEffects

SignedTransactionEffects

TransactionEffectsEnvelope

TrustedSignedTransactionEffects

UnsignedTransactionEffects

VerifiedCertifiedTransactionEffects

VerifiedSignedTransactionEffects

VerifiedTransactionEffectsEnvelope