iota_types::effects

Struct TransactionEffectsV1

pub struct TransactionEffectsV1 {
    pub status: ExecutionStatus,
    pub epoch: u64,
    pub gas_cost_summary: GasCostSummary,
    pub transaction_digest: Digest,
    pub gas_object_index: Option<u32>,
    pub events_digest: Option<Digest>,
    pub dependencies: Vec<Digest>,
    pub lamport_version: Version,
    pub changed_objects: Vec<ChangedObject>,
    pub unchanged_shared_objects: Vec<UnchangedSharedObject>,
    pub auxiliary_data_digest: Option<Digest>,
}

Version 1 of TransactionEffects

BCS

The BCS serialized form for this type is defined by the following ABNF:

transaction-effects-v1 = execution-status                   ; status
                         u64                                ; epoch
                         gas-cost-summary                   ; gas-used
                         digest                             ; transaction-digest
                         (option u32)                       ; gas-object-index
                         (option digest)                    ; events-digest
                         (vector digest)                    ; dependencies
                         u64                                ; lamport-version
                         (vector changed-object)            ; changed-objects
                         (vector unchanged-shared-object)   ; unchanged-shared-objects
                         (option digest)                    ; auxiliary-data-digest

Fields
status: ExecutionStatus

The status of the execution

epoch: u64

The epoch when this transaction was executed.

gas_cost_summary: GasCostSummary

The gas used by this transaction

transaction_digest: Digest

The transaction digest

gas_object_index: Option<u32>

The updated gas object reference, as an index into the changed_objects vector. Having a dedicated field for convenient access. System transaction that don’t require gas will leave this as None.

events_digest: Option<Digest>

The digest of the events emitted during execution, can be None if the transaction does not emit any event.

dependencies: Vec<Digest>

The set of transaction digests this transaction depends on.

lamport_version: Version

The version number of all the written Move objects by this transaction.

changed_objects: Vec<ChangedObject>

Objects whose state are changed in the object store.

unchanged_shared_objects: Vec<UnchangedSharedObject>

Shared objects that are not mutated in this transaction. Unlike owned objects, read-only shared objects’ version are not committed in the transaction, and in order for a node to catch up and execute it without consensus sequencing, the version needs to be committed in the effects.

auxiliary_data_digest: Option<Digest>

Auxiliary data that are not protocol-critical, generated as part of the effects but are stored separately. Storing it separately allows us to avoid bloating the effects with data that are not critical. It also provides more flexibility on the format and type of the data.