iota_types

Module gas_coin

Structs

GAS

GasCoin

Rust version of the Move iota::coin::CoinIota::iota::IOTA type

IotaTreasuryCap

Constants

GAS_MODULE_NAME

GAS_STRUCT_NAME

GAS_TREASURY_CAP_STRUCT_NAME

NANOS_PER_IOTA

The number of Nanos per IOTA token

STARDUST_TOTAL_SUPPLY_IOTA

Total supply in IOTA at genesis, after the migration from a Stardust ledger, before any inflation mechanism

STARDUST_TOTAL_SUPPLY_NANOS

Total supply at genesis denominated in Nanos, after the migration from a Stardust ledger, before any inflation mechanism