iota_types

Module governance

Structs

StakedIota

Constants

ADD_STAKE_FUN_NAME

ADD_STAKE_MUL_COIN_FUN_NAME

MAX_VALIDATOR_COUNT

Maximum number of active validators at any moment. We do not allow the number of validators in any epoch to go above this.

MIN_VALIDATOR_JOINING_STAKE_NANOS

Lower-bound on the amount of stake required to become a validator.

STAKED_IOTA_STRUCT_NAME

STAKING_POOL_MODULE_NAME

VALIDATOR_LOW_STAKE_GRACE_PERIOD

A validator can have stake below validator_low_stake_threshold for this many epochs before being kicked out.

VALIDATOR_LOW_STAKE_THRESHOLD_NANOS

Validators with stake amount below validator_low_stake_threshold are considered to have low stake and will be escorted out of the validator set after being below this threshold for more than validator_low_stake_grace_period number of epochs.

VALIDATOR_VERY_LOW_STAKE_THRESHOLD_NANOS

Validators with stake below validator_very_low_stake_threshold will be removed immediately at epoch change, no grace period.

WITHDRAW_STAKE_FUN_NAME