iota_types

Module iota_serde

Structs

AsProtocolVersion

BigInt

HexAccountAddress

custom serde for AccountAddress

IotaBitmap

Serializes a bitmap according to the roaring bitmap on-disk standard. https://github.com/RoaringBitmap/RoaringFormatSpec

IotaStructTag

IotaTypeTag

Readable

Use with serde_as to control serde for human-readable serialization and deserialization H : serde_as SerializeAs/DeserializeAs delegation for human readable in/output R : serde_as SerializeAs/DeserializeAs delegation for non-human readable in/output

SequenceNumber

Functions

to_iota_struct_tag_string

Serialize StructTag as a string, retaining the leading zeros in the address.