iota_types::iota_system_state

Module iota_system_state_inner_v2

Structs

IotaSystemStateV2

Rust version of the Move iota_framework::iota_system::iota_system::IotaSystemStateV2 type. An additional field safe_mode_computation_charges_burned is added over IotaSystemStateV2 to allow for burning of base fees in safe mode when protocol_defined_base_fee is enabled in the protocol config.

ValidatorSetV2

Rust version of the Move iota_system::validator_set::ValidatorSetV2 type The second version of the struct storing information about validator set. This version is an extension on the first one, that supports a new approach to committee selection, where committee members taking part in consensus are selected from a set of active_validators before an epoch begins. committee_members is a vector of indices of validators stored in active_validators, that have been selected to take part in consensus during the current epoch.