iota_types::iota_system_state

Module iota_system_state_summary

Structs

IotaSystemStateSummaryV1

This is the JSON-RPC type for the IotaSystemStateV1 object. It flattens all fields to make them top-level fields such that it as minimum dependencies to the internal data structures of the IOTA system state type.

IotaSystemStateSummaryV2

This is the JSON-RPC type for the IotaSystemStateV2 object. It flattens all fields to make them top-level fields such that it as minimum dependencies to the internal data structures of the IOTA system state type.

IotaValidatorSummary

This is the JSON-RPC type for the IOTA validator. It flattens all inner structures to top-level fields so that they are decoupled from the internal definitions.

Enums

IotaSystemStateSummary

This is the JSON-RPC type for IOTA system state objects. It is an enum type that can represent either V1 or V2 system state objects.

Functions

get_validator_by_pool_id

Given the staking pool id of a validator, return the validator’s IotaValidatorSummary, works for validator candidates, active validators, as well as inactive validators.