iota_types

Module messages_grpc

Structs

HandleCapabilityNotificationRequestV1

HandleCapabilityNotificationResponseV1

HandleCertificateRequestV1

HandleCertificateResponseV1

Response type for version 1 of the handle certificate validator API.

HandleSoftBundleCertificatesRequestV1

Soft Bundle request. See SIP-19.

HandleSoftBundleCertificatesResponseV1

Response type for the handle Soft Bundle certificates validator API. If wait_for_effects is true, it is guaranteed that:

HandleTransactionResponse

ObjectInfoRequest

A request for information about an object and optionally its parent certificate at a specific version.

ObjectInfoResponse

This message provides information about the latest object and its lock.

SubmitCertificateResponse

SystemStateRequest

TransactionInfoRequest

TransactionInfoResponse

VerifiedHandleCertificateResponse

VerifiedObjectInfoResponse

Verified version of ObjectInfoResponse. layout and lock_for_debugging are skipped because they are not needed and we don’t want to verify them.

Enums

LayoutGenerationOption

Layout generation options – you can either generate or not generate the layout.

ObjectInfoRequestKind

TransactionStatus