iota_types

Module multisig

Structs

MultiSig

Aggregated signature from members of a multisig committee.

MultiSigPublicKey

A multisig committee

MultisigMember

A member in a multisig committee

Enums

MultisigMemberSignature

A signature from a member of a multisig committee.

Type Aliases

BitmapUnit

ThresholdUnit

WeightUnit

Attribute Macros

enum_dispatch

Annotating a trait or enum definition with an #[enum_dispatch] attribute will register it with the enum_dispatch library, allowing it to be used to generate impl blocks elsewhere.