Function iota_types::passkey_authenticator::to_signing_digest

pub fn to_signing_digest<T: Serialize>(
    intent_msg: &IntentMessage<T>,
) -> [u8; 32]

Compute the BCS hash of the value in intent message. In the case of transaction data, this is the BCS hash of struct TransactionData, different from the transaction digest itself that computes the BCS hash of the Rust type prefix and struct TransactionData. (See fn digest in impl Message for SenderSignedData).