iota_types::timelock

Module timelock

Structs

TimeLock

Rust version of the Move stardust::TimeLock type.

Enums

VestedRewardError

Constants

TIMELOCK_MODULE_NAME

TIMELOCK_STRUCT_NAME

VESTED_REWARD_ID_PREFIX

All basic outputs whose IDs start with this prefix represent vested rewards that were created during the stardust upgrade on IOTA mainnet.

Functions

is_timelock

Is this other StructTag representing a TimeLock?

is_timelocked_balance

Is this other StructTag representing a TimeLock<Balance<T>>?

is_timelocked_gas_balance

Is this other StructTag representing a TimeLock<Balance<IOTA>>?

is_timelocked_vested_reward

Checks if an output is a timelocked vested reward.

is_vested_reward

Checks if an output is a vested reward, if it has a specific ID prefix, and if it contains a timelock unlock condition, and if an output has no native tokens, and if an output has only 2 unlock conditions and their address.

to_genesis_object

Creates a genesis object from a time-locked balance.

try_from_stardust

Creates a TimeLock<Balance<IOTA>> from a Stardust-based Basic Output that represents a vested reward.