iota_types

Module transaction

Structs

CheckedInputObjects

GasData

Payment information for executing a transaction

InputObjects

ObjectReadResult

The result of reading an object for execution. Because shared objects may be deleted, one possible result of reading a shared object is that ObjectReadResultKind::Deleted is returned.

ReceivingObjectReadResult

ReceivingObjects

SenderSignedData

SenderSignedTransaction

SharedObjectRef

A shared object input to a programmable transaction

SystemPackage

TransactionDataV1

Enums

InputObjectKind

ObjectReadResultKind

ReceivingObjectReadResultKind

TransactionData

Transaction

TransactionKey

TransactionKey uniquely identifies a transaction across all epochs. Note that a single transaction may have multiple keys, for example a RandomnessStateUpdate could be identified by both Digest and RandomnessRound.

Constants

DEFAULT_VALIDATOR_GAS_PRICE

GAS_PRICE_FOR_SYSTEM_TX

TEST_ONLY_GAS_UNIT_FOR_GENERIC

TEST_ONLY_GAS_UNIT_FOR_HEAVY_COMPUTATION_STORAGE

TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS

TEST_ONLY_GAS_UNIT_FOR_PUBLISH

TEST_ONLY_GAS_UNIT_FOR_SPLIT_COIN

TEST_ONLY_GAS_UNIT_FOR_STAKING

TEST_ONLY_GAS_UNIT_FOR_TRANSFER

Traits

CallArgExt

Extension trait for CallArg providing helper methods.

CommandExt

MoveCallExt

ProgrammableTransactionExt

TransactionDataAPI

API for accessing and constructing TransactionData.

TransactionKindExt

Functions

auth_digest_for_sig

Computes the auth digest for a single GenericSignature.

type_tag_validity_check

Type Aliases

CallArg

Type alias for the SDK’s Input type, used as transaction call arguments.

CertifiedTransaction

SignedTransaction

A transaction that is signed by a sender and also by an authority.

Transaction

A transaction that is signed by a sender but not yet by an authority.

TrustedCertificate

TrustedTransaction

VerifiedCertificate

VerifiedSignedTransaction

VerifiedTransaction