iota_types::transaction

Enum Command

#[non_exhaustive]
pub enum Command {
    MoveCall(MoveCall),
    TransferObjects(TransferObjects),
    SplitCoins(SplitCoins),
    MergeCoins(MergeCoins),
    Publish(Publish),
    MakeMoveVector(MakeMoveVector),
    Upgrade(Upgrade),
}

A single command in a programmable transaction.

BCS

The BCS serialized form for this type is defined by the following ABNF:

command =  command-move-call
        =/ command-transfer-objects
        =/ command-split-coins
        =/ command-merge-coins
        =/ command-publish
        =/ command-make-move-vector
        =/ command-upgrade

command-move-call           = %d00 move-call
command-transfer-objects    = %d01 transfer-objects
command-split-coins         = %d02 split-coins
command-merge-coins         = %d03 merge-coins
command-publish             = %d04 publish
command-make-move-vector    = %d05 make-move-vector
command-upgrade             = %d06 upgrade

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

MoveCall(MoveCall)

A call to either an entry or a public Move function

TransferObjects(TransferObjects)

(Vec<forall T:key+store. T>, address) It sends n-objects to the specified address. These objects must have store (public transfer) and either the previous owner must be an address or the object must be newly created.

SplitCoins(SplitCoins)

(&mut Coin<T>, Vec<u64>) -> Vec<Coin<T>> It splits off some amounts into a new coins with those amounts

MergeCoins(MergeCoins)

(&mut Coin<T>, Vec<Coin<T>>) It merges n-coins into the first coin

Publish(Publish)

Publishes a Move package. It takes the package bytes and a list of the package’s transitive dependencies to link against on-chain.

MakeMoveVector(MakeMoveVector)

forall T: Vec<T> -> vector<T> Given n-values of the same type, it constructs a vector. For non objects or an empty vector, the type tag must be specified.

Upgrade(Upgrade)

Upgrades a Move package Takes (in order):

A vector of serialized modules for the package.

A vector of object ids for the transitive dependencies of the new package.

The object ID of the package being upgraded.

An argument holding the UpgradeTicket that must have been produced from an earlier command in the same programmable transaction.

Implementations

impl Command

pub fn is_move_call(&self) -> bool

Checks if this is a move_call variant.

pub fn as_move_call(&self) -> &MoveCall

Converts this into a move_call if it is a move_call variant, or panics otherwise.

pub fn as_move_call_mut(&mut self) -> &mut MoveCall

Converts this into a mut move_call if it is a move_call variant, or panics otherwise.

pub fn as_move_call_opt(&self) -> Option<&MoveCall>

Converts this into a move_call if it is a move_call variant, or returns None otherwise.

pub fn as_move_call_mut_opt(&mut self) -> Option<&mut MoveCall>

Converts this into a mut move_call if it is a move_call variant, or returns None otherwise.

pub fn into_move_call_opt(self) -> Option<MoveCall>

Converts this into a move_call if it is a move_call variant, or returns None otherwise.

pub fn into_move_call(self) -> MoveCall

Converts this into a move_call if it is a move_call variant, or panics otherwise.

pub fn is_transfer_objects(&self) -> bool

Checks if this is a transfer_objects variant.

pub fn as_transfer_objects(&self) -> &TransferObjects

Converts this into a transfer_objects if it is a transfer_objects variant, or panics otherwise.

pub fn as_transfer_objects_mut(&mut self) -> &mut TransferObjects

Converts this into a mut transfer_objects if it is a transfer_objects variant, or panics otherwise.

pub fn as_transfer_objects_opt(&self) -> Option<&TransferObjects>

Converts this into a transfer_objects if it is a transfer_objects variant, or returns None otherwise.

pub fn as_transfer_objects_mut_opt(&mut self) -> Option<&mut TransferObjects>

Converts this into a mut transfer_objects if it is a transfer_objects variant, or returns None otherwise.

pub fn into_transfer_objects_opt(self) -> Option<TransferObjects>

Converts this into a transfer_objects if it is a transfer_objects variant, or returns None otherwise.

pub fn into_transfer_objects(self) -> TransferObjects

Converts this into a transfer_objects if it is a transfer_objects variant, or panics otherwise.

pub fn is_split_coins(&self) -> bool

Checks if this is a split_coins variant.

pub fn as_split_coins(&self) -> &SplitCoins

Converts this into a split_coins if it is a split_coins variant, or panics otherwise.

pub fn as_split_coins_mut(&mut self) -> &mut SplitCoins

Converts this into a mut split_coins if it is a split_coins variant, or panics otherwise.

pub fn as_split_coins_opt(&self) -> Option<&SplitCoins>

Converts this into a split_coins if it is a split_coins variant, or returns None otherwise.

pub fn as_split_coins_mut_opt(&mut self) -> Option<&mut SplitCoins>

Converts this into a mut split_coins if it is a split_coins variant, or returns None otherwise.

pub fn into_split_coins_opt(self) -> Option<SplitCoins>

Converts this into a split_coins if it is a split_coins variant, or returns None otherwise.

pub fn into_split_coins(self) -> SplitCoins

Converts this into a split_coins if it is a split_coins variant, or panics otherwise.

pub fn is_merge_coins(&self) -> bool

Checks if this is a merge_coins variant.

pub fn as_merge_coins(&self) -> &MergeCoins

Converts this into a merge_coins if it is a merge_coins variant, or panics otherwise.

pub fn as_merge_coins_mut(&mut self) -> &mut MergeCoins

Converts this into a mut merge_coins if it is a merge_coins variant, or panics otherwise.

pub fn as_merge_coins_opt(&self) -> Option<&MergeCoins>

Converts this into a merge_coins if it is a merge_coins variant, or returns None otherwise.

pub fn as_merge_coins_mut_opt(&mut self) -> Option<&mut MergeCoins>

Converts this into a mut merge_coins if it is a merge_coins variant, or returns None otherwise.

pub fn into_merge_coins_opt(self) -> Option<MergeCoins>

Converts this into a merge_coins if it is a merge_coins variant, or returns None otherwise.

pub fn into_merge_coins(self) -> MergeCoins

Converts this into a merge_coins if it is a merge_coins variant, or panics otherwise.

pub fn is_publish(&self) -> bool

Checks if this is a publish variant.

pub fn as_publish(&self) -> &Publish

Converts this into a publish if it is a publish variant, or panics otherwise.

pub fn as_publish_mut(&mut self) -> &mut Publish

Converts this into a mut publish if it is a publish variant, or panics otherwise.

pub fn as_publish_opt(&self) -> Option<&Publish>

Converts this into a publish if it is a publish variant, or returns None otherwise.

pub fn as_publish_mut_opt(&mut self) -> Option<&mut Publish>

Converts this into a mut publish if it is a publish variant, or returns None otherwise.

pub fn into_publish_opt(self) -> Option<Publish>

Converts this into a publish if it is a publish variant, or returns None otherwise.

pub fn into_publish(self) -> Publish

Converts this into a publish if it is a publish variant, or panics otherwise.

pub fn is_make_move_vector(&self) -> bool

Checks if this is a make_move_vector variant.

pub fn as_make_move_vector(&self) -> &MakeMoveVector

Converts this into a make_move_vector if it is a make_move_vector variant, or panics otherwise.

pub fn as_make_move_vector_mut(&mut self) -> &mut MakeMoveVector

Converts this into a mut make_move_vector if it is a make_move_vector variant, or panics otherwise.

pub fn as_make_move_vector_opt(&self) -> Option<&MakeMoveVector>

Converts this into a make_move_vector if it is a make_move_vector variant, or returns None otherwise.

pub fn as_make_move_vector_mut_opt(&mut self) -> Option<&mut MakeMoveVector>

Converts this into a mut make_move_vector if it is a make_move_vector variant, or returns None otherwise.

pub fn into_make_move_vector_opt(self) -> Option<MakeMoveVector>

Converts this into a make_move_vector if it is a make_move_vector variant, or returns None otherwise.

pub fn into_make_move_vector(self) -> MakeMoveVector

Converts this into a make_move_vector if it is a make_move_vector variant, or panics otherwise.

pub fn is_upgrade(&self) -> bool

Checks if this is a upgrade variant.

pub fn as_upgrade(&self) -> &Upgrade

Converts this into a upgrade if it is a upgrade variant, or panics otherwise.

pub fn as_upgrade_mut(&mut self) -> &mut Upgrade

Converts this into a mut upgrade if it is a upgrade variant, or panics otherwise.

pub fn as_upgrade_opt(&self) -> Option<&Upgrade>

Converts this into a upgrade if it is a upgrade variant, or returns None otherwise.

pub fn as_upgrade_mut_opt(&mut self) -> Option<&mut Upgrade>

Converts this into a mut upgrade if it is a upgrade variant, or returns None otherwise.

pub fn into_upgrade_opt(self) -> Option<Upgrade>

Converts this into a upgrade if it is a upgrade variant, or returns None otherwise.

pub fn into_upgrade(self) -> Upgrade

Converts this into a upgrade if it is a upgrade variant, or panics otherwise.

pub fn new_move_call( package: ObjectId, module: Identifier, function: Identifier, type_arguments: Vec<TypeTag>, arguments: Vec<Argument>, ) -> Command

Create a command to call a Move function.

pub fn new_transfer_objects( objects: Vec<Argument>, address: Argument, ) -> Command

Create a command to transfer objects to an address.

pub fn new_split_coins(coin: Argument, amounts: Vec<Argument>) -> Command

Create a command to split a coin into multiple coins by amounts.

pub fn new_merge_coins(coin: Argument, coins_to_merge: Vec<Argument>) -> Command

Create a command to merge multiple coins into one.

pub fn new_publish( modules: Vec<Vec<u8>>, dependencies: Vec<ObjectId>, ) -> Command

Create a command to publish a new Move package.

pub fn new_make_move_vector( type_: Option<TypeTag>, elements: Vec<Argument>, ) -> Command

Create a command to construct a Move vector from elements.

pub fn new_upgrade( modules: Vec<Vec<u8>>, dependencies: Vec<ObjectId>, package: ObjectId, ticket: Argument, ) -> Command

Create a command to upgrade an existing Move package.