iota_types::transaction

Trait CallArgExt

pub trait CallArgExt: Sized + Sealed {
    // Required methods
    fn input_object_kind(&self) -> Option<InputObjectKind>;
    fn validity_check(&self, config: &ProtocolConfig) -> UserInputResult;
}

Extension trait for CallArg providing helper methods.

Required Methods

fn input_object_kind(&self) -> Option<InputObjectKind>

Returns the input object kind for this argument, excluding receiving objects.

fn validity_check(&self, config: &ProtocolConfig) -> UserInputResult

Validity check for this argument against the given protocol config.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.