iota_types::transaction

Type Alias CallArg

pub type CallArg = Input;

Type alias for the SDK’s Input type, used as transaction call arguments.

Aliased Type

pub enum CallArg {
    Pure(Vec<u8>),
    ImmutableOrOwned(ObjectReference),
    Shared(SharedObjectReference),
    Receiving(ObjectReference),
}

Variants

Pure(Vec<u8>)

A move value serialized as BCS.

For normal operations this is required to be a move primitive type and not contain structs or objects.

ImmutableOrOwned(ObjectReference)

A move object that is either immutable or address owned

Shared(SharedObjectReference)

A move object whose owner is “Shared”

Receiving(ObjectReference)

A move object that is attempted to be received in this transaction.