Module iota_types::base_types

Re-exports

pub use crate::committee::EpochId;

pub use crate::digests::ObjectDigest;

pub use crate::digests::TransactionDigest;

pub use crate::digests::TransactionEffectsDigest;

Structs

ExecutionData

ExecutionDigests

IotaAddress

MoveObjectType

Wrapper around StructTag with a space-efficient representation for common types like coins The StructTag for a gas coin is 84 bytes, so using 1 byte instead is a win. The inner representation is private to prevent incorrectly constructing an Other instead of one of the specialized variants, e.g. Other(GasCoin::type_()) instead of GasCoin

ObjectID

ObjectInfo

SequenceNumber

SizeOneVec

TxContext

UserData

VerifiedExecutionData

Enums

MoveObjectType_

Even though it is declared public, it is the “private”, internal representation for MoveObjectType

ObjectIDParseError

ObjectType

Type of a Iota object

TxContextKind

Constants

IOTA_ADDRESS_LENGTH

RESOLVED_ASCII_STR

RESOLVED_STD_OPTION

RESOLVED_UTF8_STR

STD_ASCII_MODULE_NAME

STD_ASCII_STRUCT_NAME

STD_OPTION_MODULE_NAME

STD_OPTION_STRUCT_NAME

STD_UTF8_MODULE_NAME

STD_UTF8_STRUCT_NAME

TX_CONTEXT_MODULE_NAME

TX_CONTEXT_STRUCT_NAME

Traits

ConciseableName

Functions

dbg_addr

Generate a fake IotaAddress with repeated one byte.

dbg_object_id

Generate a fake ObjectID with repeated one byte.

is_primitive_type_tag

Whether this type is valid as a primitive (pure) transaction input.

random_object_ref

update_object_ref_for_testing

Type Aliases

AuthorityName

CommitRound

The round number.

ObjectRef

TxSequenceNumber

VersionDigest

VersionNumber