iota_types

Module coin_manager

Structs

CoinManager

The purpose of a CoinManager is to allow access to all properties of a Coin on-chain from within a single shared object This includes access to the total supply and metadata In addition a optional maximum supply can be set and a custom additional Metadata field can be added. Holds all related objects to a Coin in a convenient shared function.

CoinManagerTreasuryCap

Like TreasuryCap, but for dealing with TreasuryCap inside CoinManager objects

ImmutableCoinMetadata

The immutable version of CoinMetadata, used in case of migrating from frozen objects to a CoinManager holding the metadata.

Constants

COIN_MANAGER_MODULE_NAME

COIN_MANAGER_STRUCT_NAME

COIN_MANAGER_TREASURY_CAP_STRUCT_NAME