Module iota_types::digests

Structs

CertificateDigest

A digest of a certificate, which commits to the signatures as well as the tx.

ChainIdentifier

Representation of a network’s identifier by the genesis checkpoint’s digest

CheckpointContentsDigest

CheckpointDigest

Representation of a Checkpoint’s digest

ConsensusCommitDigest

Digest

A representation of a 32 byte digest

EffectsAuxDataDigest

ObjectDigest

SenderSignedDataDigest

A digest of a SenderSignedData, which commits to the signatures as well as the tx.

TransactionDigest

A transaction will have a (unique) digest.

TransactionEffectsDigest

TransactionEventsDigest

ZKLoginInputsDigest

A digest of a ZkLoginInputs, which commits to the signatures as well as the tx.

Statics

MAINNET_CHAIN_IDENTIFIER

TESTNET_CHAIN_IDENTIFIER

Functions

get_mainnet_chain_identifier

get_testnet_chain_identifier