iota_types

Module effects

Structs

EffectsObjectChange

ObjectChange

TestEffectsBuilder

TransactionEffectsDebugSummary

TransactionEvents

Enums

IDOperation

InputSharedObject

ObjectIn

If an object exists (at root-level) in the store prior to this transaction, it should be Exist, otherwise it’s NonExist, e.g. wrapped objects should be NonExist.

ObjectOut

ObjectRemoveKind

TransactionEffects

The response from processing a transaction or a certified transaction

UnchangedSharedKind

Constants

APPROX_SIZE_OF_EPOCH_ID

APPROX_SIZE_OF_EXECUTION_STATUS

APPROX_SIZE_OF_GAS_COST_SUMMARY

APPROX_SIZE_OF_OBJECT_REF

APPROX_SIZE_OF_OPT_TX_EVENTS_DIGEST

APPROX_SIZE_OF_OWNER

APPROX_SIZE_OF_TX_DIGEST

Traits

TransactionEffectsAPI

Type Aliases

CertifiedTransactionEffects

SignedTransactionEffects

TransactionEffectsEnvelope

TrustedSignedTransactionEffects

UnsignedTransactionEffects

VerifiedCertifiedTransactionEffects

VerifiedSignedTransactionEffects

VerifiedTransactionEffectsEnvelope