iota_types

Module event

Structs

Event

An event

EventEnvelope

A universal IOTA event type encapsulating different types of events

EventID

Unique ID of an IOTA Event, the ID is a combination of transaction digest and event seq number.

SystemEpochInfoEventV1

Event emitted in move code fun advance_epoch in protocol versions 1 to 3

SystemEpochInfoEventV2

Event emitted in move code fun advance_epoch in protocol versions 5 and later. This second version of the event includes the tips amount to show how much of the gas fees go to the validators when protocol_defined_base_fee is enabled in the protocol config.

Enums

SystemEpochInfoEvent