iota_types

Module iota_serde

Structs

AsProtocolVersion

BigInt

IotaStructTag

IotaTypeTag

Readable

Use with serde_as to control serde for human-readable serialization and deserialization H : serde_as SerializeAs/DeserializeAs delegation for human readable in/output R : serde_as SerializeAs/DeserializeAs delegation for non-human readable in/output

SequenceNumber

TypeName

A marker for type tags that are serialized as strings. Normally, a type tag is serialized as a string for readable formats, and as a byte array for non-readable formats. This marker can be used to serialize a type tag as a string even in non-readable formats.

Functions

to_iota_struct_tag_string

Serialize StructTag as a string, retaining the leading zeros in the address.