iota_types::iota_system_state

Module epoch_start_iota_system_state

Structs

EpochStartSystemStateV1

EpochStartSystemStateV2

EpochStartValidatorInfoV1

Enums

EpochStartSystemState

This type captures the minimum amount of information from IotaSystemState needed by a validator to run the protocol. This allows us to decouple from the actual IotaSystemState type, and hence do not need to evolve it when we upgrade the IotaSystemState type. Evolving EpochStartSystemState is also a lot easier in that we could add optional fields and fill them with None for older versions. When we absolutely must delete fields, we could also add new db tables to store the new version. This is OK because we only store one copy of this as part of EpochStartConfiguration for the most recent epoch in the db.

Traits

EpochStartSystemStateTrait

Functions

convert_validator_to_epoch_start_info

Converts a validator into EpochStartValidatorInfoV1