iota_types

Module messages_consensus

Structs

AuthorityCapabilitiesDigest

AuthorityCapabilitiesV1

Used to advertise capabilities of each authority via consensus. This allows validators to negotiate the creation of the ChangeEpoch transaction.

ConsensusCommitPrologueV1

ConsensusTransaction

Enums

ConsensusDeterminedVersionAssignments

Uses an enum to allow for future expansion of the ConsensusDeterminedVersionAssignments.

ConsensusTransactionKey

ConsensusTransactionKind

VersionedDkgConfirmation

VersionedDkgMessage

Functions

check_total_jwk_size

Type Aliases

SignedAuthorityCapabilitiesV1

TimestampMs

Non-decreasing timestamp produced by consensus in ms.

VerifiedAuthorityCapabilitiesV1