iota_types::move_authenticator

Trait MoveAuthenticatorExt

pub trait MoveAuthenticatorExt: Sized + Sealed {
    // Required methods
    fn address(&self) -> Address;
    fn call_args(&self) -> &[CallArg] ;
    fn type_args(&self) -> &[TypeTag];
    fn object_to_authenticate(&self) -> &CallArg;
    fn object_to_authenticate_components(
        &self,
    ) -> UserInputResult<(ObjectId, Option<Version>, Option<ObjectDigest>)>;
    fn input_objects(&self) -> Vec<InputObjectKind>;
    fn receiving_objects(&self) -> Vec<ObjectReference>;
    fn shared_objects(&self) -> Vec<SharedObjectReference>;
    fn validity_check(&self, config: &ProtocolConfig) -> UserInputResult;
}

Required Methods

fn address(&self) -> Address

Returns the address of the object being authenticated, which acts as the sender of the transaction.

fn call_args(&self) -> &[CallArg]

Returns the input objects or primitive values passed to the authenticate function.

fn type_args(&self) -> &[TypeTag]

Returns the type arguments for the Move authenticate function.

fn object_to_authenticate(&self) -> &CallArg

Returns the object that is being authenticated (the account/sender).

fn object_to_authenticate_components( &self, ) -> UserInputResult<(ObjectId, Option<Version>, Option<ObjectDigest>)>

fn input_objects(&self) -> Vec<InputObjectKind>

fn receiving_objects(&self) -> Vec<ObjectReference>

fn shared_objects(&self) -> Vec<SharedObjectReference>

fn validity_check(&self, config: &ProtocolConfig) -> UserInputResult

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".