Module iota_types::quorum_driver_types

Structs

ExecuteTransactionRequestV1

ExecuteTransactionResponseV1

FinalizedEffects

QuorumDriverResponse

Enums

EffectsFinalityInfo

ExecuteTransactionRequestType

QuorumDriverError

Client facing errors regarding transaction submission via Quorum Driver. Every invariant needs detailed documents to instruct client handling.

Constants

NON_RECOVERABLE_ERROR_MSG

Type Aliases

GroupedErrors

IsTransactionExecutedLocally

When requested to execute a transaction with WaitForLocalExecution, TransactionOrchestrator attempts to execute this transaction locally after it is finalized. This value represents whether the transaction is confirmed to be executed on this node before the response returns.

QuorumDriverEffectsQueueResult

QuorumDriverResult