iota_types::storage

Trait ReadStore

pub trait ReadStore: ObjectStore {

    // Required methods
    fn try_get_committee(
        &self,
        epoch: EpochId,
    ) -> Result<Option<Arc<Committee>>>;
    fn try_get_latest_checkpoint(&self) -> Result<VerifiedCheckpoint>;
    fn try_get_highest_verified_checkpoint(&self) -> Result<VerifiedCheckpoint>;
    fn try_get_highest_synced_checkpoint(&self) -> Result<VerifiedCheckpoint>;
    fn try_get_lowest_available_checkpoint(
        &self,
    ) -> Result<CheckpointSequenceNumber>;
    fn try_get_checkpoint_by_digest(
        &self,
        digest: &CheckpointDigest,
    ) -> Result<Option<VerifiedCheckpoint>>;
    fn try_get_checkpoint_by_sequence_number(
        &self,
        sequence_number: CheckpointSequenceNumber,
    ) -> Result<Option<VerifiedCheckpoint>>;
    fn try_get_checkpoint_contents_by_digest(
        &self,
        digest: &CheckpointContentsDigest,
    ) -> Result<Option<CheckpointContents>>;
    fn try_get_checkpoint_contents_by_sequence_number(
        &self,
        sequence_number: CheckpointSequenceNumber,
    ) -> Result<Option<CheckpointContents>>;
    fn try_get_transaction(
        &self,
        tx_digest: &TransactionDigest,
    ) -> Result<Option<Arc<VerifiedTransaction>>>;
    fn try_get_transaction_effects(
        &self,
        tx_digest: &TransactionDigest,
    ) -> Result<Option<TransactionEffects>>;
    fn try_get_events(
        &self,
        digest: &TransactionDigest,
    ) -> Result<Option<TransactionEvents>>;
    fn try_get_full_checkpoint_contents_by_sequence_number(
        &self,
        sequence_number: CheckpointSequenceNumber,
    ) -> Result<Option<FullCheckpointContents>>;
    fn try_get_full_checkpoint_contents(
        &self,
        digest: &CheckpointContentsDigest,
    ) -> Result<Option<FullCheckpointContents>>;

    // Provided methods
    fn get_committee(&self, epoch: EpochId) -> Option<Arc<Committee>> { ... }
    fn get_latest_checkpoint(&self) -> VerifiedCheckpoint { ... }
    fn try_get_latest_checkpoint_sequence_number(
        &self,
    ) -> Result<CheckpointSequenceNumber> { ... }
    fn get_latest_checkpoint_sequence_number(&self) -> CheckpointSequenceNumber { ... }
    fn try_get_latest_epoch_id(&self) -> Result<EpochId> { ... }
    fn get_latest_epoch_id(&self) -> EpochId { ... }
    fn get_highest_verified_checkpoint(&self) -> VerifiedCheckpoint { ... }
    fn get_highest_synced_checkpoint(&self) -> VerifiedCheckpoint { ... }
    fn get_lowest_available_checkpoint(&self) -> CheckpointSequenceNumber { ... }
    fn get_checkpoint_by_digest(
        &self,
        digest: &CheckpointDigest,
    ) -> Option<VerifiedCheckpoint> { ... }
    fn get_checkpoint_by_sequence_number(
        &self,
        sequence_number: CheckpointSequenceNumber,
    ) -> Option<VerifiedCheckpoint> { ... }
    fn get_checkpoint_contents_by_digest(
        &self,
        digest: &CheckpointContentsDigest,
    ) -> Option<CheckpointContents> { ... }
    fn get_checkpoint_contents_by_sequence_number(
        &self,
        sequence_number: CheckpointSequenceNumber,
    ) -> Option<CheckpointContents> { ... }
    fn get_transaction(
        &self,
        tx_digest: &TransactionDigest,
    ) -> Option<Arc<VerifiedTransaction>> { ... }
    fn try_multi_get_transactions(
        &self,
        tx_digests: &[TransactionDigest],
    ) -> Result<Vec<Option<Arc<VerifiedTransaction>>>> { ... }
    fn multi_get_transactions(
        &self,
        tx_digests: &[TransactionDigest],
    ) -> Vec<Option<Arc<VerifiedTransaction>>> { ... }
    fn get_transaction_effects(
        &self,
        tx_digest: &TransactionDigest,
    ) -> Option<TransactionEffects> { ... }
    fn try_multi_get_transaction_effects(
        &self,
        tx_digests: &[TransactionDigest],
    ) -> Result<Vec<Option<TransactionEffects>>> { ... }
    fn multi_get_transaction_effects(
        &self,
        tx_digests: &[TransactionDigest],
    ) -> Vec<Option<TransactionEffects>> { ... }
    fn get_events(
        &self,
        digest: &TransactionDigest,
    ) -> Option<TransactionEvents> { ... }
    fn try_multi_get_events(
        &self,
        digests: &[TransactionDigest],
    ) -> Result<Vec<Option<TransactionEvents>>> { ... }
    fn multi_get_events(
        &self,
        digests: &[TransactionDigest],
    ) -> Vec<Option<TransactionEvents>> { ... }
    fn get_full_checkpoint_contents_by_sequence_number(
        &self,
        sequence_number: CheckpointSequenceNumber,
    ) -> Option<FullCheckpointContents> { ... }
    fn get_full_checkpoint_contents(
        &self,
        digest: &CheckpointContentsDigest,
    ) -> Option<FullCheckpointContents> { ... }
    fn try_multi_get_transactions_with_events_and_effects(
        &self,
        transaction_digests: Vec<TransactionDigest>,
    ) -> Result<Vec<TransactionWithEffectsAndEvents>> { ... }
    fn get_checkpoint_transaction(
        &self,
        tx_with_events_and_effects: TransactionWithEffectsAndEvents,
    ) -> Result<CheckpointTransaction> { ... }
    fn stream_checkpoint_transactions(
        &self,
        checkpoint_contents: CheckpointContents,
    ) -> Pin<Box<dyn Stream<Item = Result<CheckpointTransaction>> + Send + '_>>
       where Self: Sync { ... }
    fn try_get_checkpoint_data(
        &self,
        checkpoint: VerifiedCheckpoint,
        checkpoint_contents: CheckpointContents,
    ) -> Result<CheckpointData> { ... }
    fn get_checkpoint_data(
        &self,
        checkpoint: VerifiedCheckpoint,
        checkpoint_contents: CheckpointContents,
    ) -> CheckpointData { ... }

}

Required Methods

fn try_get_committee(&self, epoch: EpochId) -> Result<Option<Arc<Committee>>>

fn try_get_latest_checkpoint(&self) -> Result<VerifiedCheckpoint>

Get the latest available checkpoint. This is the latest executed checkpoint.

All transactions, effects, objects and events are guaranteed to be available for the returned checkpoint.

fn try_get_highest_verified_checkpoint(&self) -> Result<VerifiedCheckpoint>

Get the highest verified checkpoint. This is the highest checkpoint summary that has been verified, generally by state-sync. Only the checkpoint header is guaranteed to be present in the store.

fn try_get_highest_synced_checkpoint(&self) -> Result<VerifiedCheckpoint>

Get the highest synced checkpoint. This is the highest checkpoint that has been synced from state-synce. The checkpoint header, contents, transactions, and effects of this checkpoint are guaranteed to be present in the store

fn try_get_lowest_available_checkpoint( &self, ) -> Result<CheckpointSequenceNumber>

Lowest available checkpoint for which transaction and checkpoint data can be requested.

Specifically this is the lowest checkpoint for which the following data can be requested:

checkpoints

transactions

effects

events

For object availability see get_lowest_available_checkpoint_objects.

fn try_get_checkpoint_by_digest( &self, digest: &CheckpointDigest, ) -> Result<Option<VerifiedCheckpoint>>

fn try_get_checkpoint_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Result<Option<VerifiedCheckpoint>>

fn try_get_checkpoint_contents_by_digest( &self, digest: &CheckpointContentsDigest, ) -> Result<Option<CheckpointContents>>

fn try_get_checkpoint_contents_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Result<Option<CheckpointContents>>

fn try_get_transaction( &self, tx_digest: &TransactionDigest, ) -> Result<Option<Arc<VerifiedTransaction>>>

fn try_get_transaction_effects( &self, tx_digest: &TransactionDigest, ) -> Result<Option<TransactionEffects>>

fn try_get_events( &self, digest: &TransactionDigest, ) -> Result<Option<TransactionEvents>>

fn try_get_full_checkpoint_contents_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Result<Option<FullCheckpointContents>>

Get a “full” checkpoint for purposes of state-sync “full” checkpoints include: header, contents, transactions, effects

fn try_get_full_checkpoint_contents( &self, digest: &CheckpointContentsDigest, ) -> Result<Option<FullCheckpointContents>>

Get a “full” checkpoint for purposes of state-sync “full” checkpoints include: header, contents, transactions, effects

Provided Methods

fn get_committee(&self, epoch: EpochId) -> Option<Arc<Committee>>

Non-fallible version of try_get_committee.

fn get_latest_checkpoint(&self) -> VerifiedCheckpoint

Non-fallible version of try_get_latest_checkpoint.

fn try_get_latest_checkpoint_sequence_number( &self, ) -> Result<CheckpointSequenceNumber>

Get the latest available checkpoint sequence number. This is the sequence number of the latest executed checkpoint.

fn get_latest_checkpoint_sequence_number(&self) -> CheckpointSequenceNumber

Non-fallible version of try_get_latest_checkpoint_sequence_number.

fn try_get_latest_epoch_id(&self) -> Result<EpochId>

Get the epoch of the latest checkpoint

fn get_latest_epoch_id(&self) -> EpochId

Non-fallible version of try_get_latest_epoch_id.

fn get_highest_verified_checkpoint(&self) -> VerifiedCheckpoint

Non-fallible version of try_get_highest_verified_checkpoint.

fn get_highest_synced_checkpoint(&self) -> VerifiedCheckpoint

Non-fallible version of try_get_highest_synced_checkpoint.

fn get_lowest_available_checkpoint(&self) -> CheckpointSequenceNumber

Non-fallible version of try_get_lowest_available_checkpoint.

fn get_checkpoint_by_digest( &self, digest: &CheckpointDigest, ) -> Option<VerifiedCheckpoint>

Non-fallible version of try_get_checkpoint_by_digest.

fn get_checkpoint_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Option<VerifiedCheckpoint>

Non-fallible version of try_get_checkpoint_by_sequence_number.

fn get_checkpoint_contents_by_digest( &self, digest: &CheckpointContentsDigest, ) -> Option<CheckpointContents>

Non-fallible version of try_get_checkpoint_contents_by_digest.

fn get_checkpoint_contents_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Option<CheckpointContents>

Non-fallible version of try_get_checkpoint_contents_by_sequence_number.

fn get_transaction( &self, tx_digest: &TransactionDigest, ) -> Option<Arc<VerifiedTransaction>>

Non-fallible version of try_get_transaction.

fn try_multi_get_transactions( &self, tx_digests: &[TransactionDigest], ) -> Result<Vec<Option<Arc<VerifiedTransaction>>>>

fn multi_get_transactions( &self, tx_digests: &[TransactionDigest], ) -> Vec<Option<Arc<VerifiedTransaction>>>

Non-fallible version of try_multi_get_transactions.

fn get_transaction_effects( &self, tx_digest: &TransactionDigest, ) -> Option<TransactionEffects>

Non-fallible version of try_get_transaction_effects.

fn try_multi_get_transaction_effects( &self, tx_digests: &[TransactionDigest], ) -> Result<Vec<Option<TransactionEffects>>>

fn multi_get_transaction_effects( &self, tx_digests: &[TransactionDigest], ) -> Vec<Option<TransactionEffects>>

Non-fallible version of try_multi_get_transaction_effects.

fn get_events(&self, digest: &TransactionDigest) -> Option<TransactionEvents>

Non-fallible version of try_get_events.

fn try_multi_get_events( &self, digests: &[TransactionDigest], ) -> Result<Vec<Option<TransactionEvents>>>

fn multi_get_events( &self, digests: &[TransactionDigest], ) -> Vec<Option<TransactionEvents>>

Non-fallible version of try_multi_get_events.

fn get_full_checkpoint_contents_by_sequence_number( &self, sequence_number: CheckpointSequenceNumber, ) -> Option<FullCheckpointContents>

Non-fallible version of try_get_full_checkpoint_contents_by_sequence_number.

fn get_full_checkpoint_contents( &self, digest: &CheckpointContentsDigest, ) -> Option<FullCheckpointContents>

Non-fallible version of try_get_full_checkpoint_contents.

fn try_multi_get_transactions_with_events_and_effects( &self, transaction_digests: Vec<TransactionDigest>, ) -> Result<Vec<TransactionWithEffectsAndEvents>>

fn get_checkpoint_transaction( &self, tx_with_events_and_effects: TransactionWithEffectsAndEvents, ) -> Result<CheckpointTransaction>

fn stream_checkpoint_transactions( &self, checkpoint_contents: CheckpointContents, ) -> Pin<Box<dyn Stream<Item = Result<CheckpointTransaction>> + Send + '_>>
where Self: Sync,

Stream checkpoint transactions individually to avoid large memory footprint. Returns a stream of individual CheckpointTransaction items along with metadata

fn try_get_checkpoint_data( &self, checkpoint: VerifiedCheckpoint, checkpoint_contents: CheckpointContents, ) -> Result<CheckpointData>

fn get_checkpoint_data( &self, checkpoint: VerifiedCheckpoint, checkpoint_contents: CheckpointContents, ) -> CheckpointData

Non-fallible version of try_get_checkpoint_data.