iota_types::timelock

Module timelock

Structs

TimeLock

Rust version of the Move stardust::TimeLock type.

Constants

VESTED_REWARD_ID_PREFIX

All basic outputs whose IDs start with this prefix represent vested rewards that were created during the stardust upgrade on IOTA mainnet.

Functions

is_timelocked_balance

Is this other StructTag representing a TimeLock<Balance<T>>?

is_timelocked_gas_balance

Is this other StructTag representing a TimeLock<Balance<IOTA>>?