iota_types

Module transaction

Structs

ChangeEpoch

System transaction used to change the epoch

ChangeEpochV2

System transaction used to change the epoch

ChangeEpochV3

ChangeEpochV4

CheckedInputObjects

GasData

Payment information for executing a transaction

GenesisObject

An object part of the initial chain state

GenesisTransaction

The genesis transaction

InputObjects

MakeMoveVector

Command to build a move vector out of a set of individual elements

MergeCoins

Command to merge multiple coins of the same type into a single coin

ObjectReadResult

The result of reading an object for execution. Because shared objects may be deleted, one possible result of reading a shared object is that ObjectReadResultKind::Deleted is returned.

ProgrammableMoveCall

Command to call a move function

ProgrammableTransaction

A user transaction

Publish

Command to publish a new move package

RandomnessStateUpdate

Randomness update

ReceivingObjectReadResult

ReceivingObjects

SenderSignedData

SenderSignedTransaction

SharedObjectRef

A shared object input to a programmable transaction

SplitCoins

Command to split a single coin object into multiple coins

SystemPackage

TransactionDataV1

TransferObjects

Command to transfer ownership of a set of objects to an address

Upgrade

Command to upgrade an already published package

Enums

Argument

An argument to a programmable transaction command

Command

A single command in a programmable transaction.

EndOfEpochTransactionKind

Operation run at the end of an epoch

InputObjectKind

ObjectReadResultKind

ReceivingObjectReadResultKind

TransactionData

Transaction

TransactionExpiration

A TTL for a transaction

TransactionKey

TransactionKey uniquely identifies a transaction across all epochs. Note that a single transaction may have multiple keys, for example a RandomnessStateUpdate could be identified by both Digest and RandomnessRound.

TransactionKind

Transaction type

Constants

DEFAULT_VALIDATOR_GAS_PRICE

GAS_PRICE_FOR_SYSTEM_TX

TEST_ONLY_GAS_UNIT_FOR_GENERIC

TEST_ONLY_GAS_UNIT_FOR_HEAVY_COMPUTATION_STORAGE

TEST_ONLY_GAS_UNIT_FOR_OBJECT_BASICS

TEST_ONLY_GAS_UNIT_FOR_PUBLISH

TEST_ONLY_GAS_UNIT_FOR_SPLIT_COIN

TEST_ONLY_GAS_UNIT_FOR_STAKING

TEST_ONLY_GAS_UNIT_FOR_TRANSFER

Traits

CallArgExt

Extension trait for CallArg providing helper methods.

CommandExt

ProgrammableMoveCallExt

ProgrammableTransactionExt

TransactionDataAPI

API for accessing and constructing TransactionData.

TransactionKindExt

Functions

type_tag_validity_check

Type Aliases

CallArg

Type alias for the SDK’s Input type, used as transaction call arguments.

CertifiedTransaction

SignedTransaction

A transaction that is signed by a sender and also by an authority.

Transaction

A transaction that is signed by a sender but not yet by an authority.

TrustedCertificate

TrustedTransaction

VerifiedCertificate

VerifiedSignedTransaction

VerifiedTransaction