iota_types::transaction

Trait TransactionDataAPI

pub trait TransactionDataAPI {

    // Required methods
    fn sender(&self) -> IotaAddress;
    fn kind(&self) -> &TransactionKind;
    fn kind_mut(&mut self) -> &mut TransactionKind;
    fn into_kind(self) -> TransactionKind;
    fn signers(&self) -> NonEmpty<IotaAddress>;
    fn gas_data(&self) -> &GasData;
    fn gas_owner(&self) -> IotaAddress;
    fn gas(&self) -> &[ObjectRef] ;
    fn gas_price(&self) -> u64;
    fn gas_budget(&self) -> u64;
    fn expiration(&self) -> &TransactionExpiration;
    fn contains_shared_object(&self) -> bool;
    fn shared_input_objects(&self) -> Vec<SharedInputObject>;
    fn move_calls(&self) -> Vec<(&ObjectID, &str, &str)>;
    fn input_objects(&self) -> UserInputResult<Vec<InputObjectKind>>;
    fn receiving_objects(&self) -> Vec<ObjectRef> ;
    fn validity_check(&self, config: &ProtocolConfig) -> UserInputResult;
    fn validity_check_no_gas_check(
        &self,
        config: &ProtocolConfig,
    ) -> UserInputResult;
    fn check_sponsorship(&self) -> UserInputResult;
    fn is_system_tx(&self) -> bool;
    fn is_genesis_tx(&self) -> bool;
    fn is_end_of_epoch_tx(&self) -> bool;
    fn is_sponsored_tx(&self) -> bool;
    fn sender_mut_for_testing(&mut self) -> &mut IotaAddress;
    fn gas_data_mut(&mut self) -> &mut GasData;
    fn expiration_mut_for_testing(&mut self) -> &mut TransactionExpiration;

}

Required Methods

fn sender(&self) -> IotaAddress

fn kind(&self) -> &TransactionKind

fn kind_mut(&mut self) -> &mut TransactionKind

fn into_kind(self) -> TransactionKind

fn signers(&self) -> NonEmpty<IotaAddress>

Transaction signer and Gas owner

fn gas_data(&self) -> &GasData

fn gas_owner(&self) -> IotaAddress

fn gas(&self) -> &[ObjectRef]

fn gas_price(&self) -> u64

fn gas_budget(&self) -> u64

fn expiration(&self) -> &TransactionExpiration

fn contains_shared_object(&self) -> bool

Checks if the transaction data contains at least one shared object.

IMPORTANT: This function does not check shared objects associated with MoveAuthenticator signatures. To check those objects as well, use the corresponding function from SenderSignedData.

fn shared_input_objects(&self) -> Vec<SharedInputObject>

Returns a list of the transaction data shared input objects.

IMPORTANT: This function does not return shared objects associated with MoveAuthenticator signatures. To check those objects as well, use the corresponding function from SenderSignedData.

fn move_calls(&self) -> Vec<(&ObjectID, &str, &str)>

fn input_objects(&self) -> UserInputResult<Vec<InputObjectKind>>

fn receiving_objects(&self) -> Vec<ObjectRef>

fn validity_check(&self, config: &ProtocolConfig) -> UserInputResult

fn validity_check_no_gas_check( &self, config: &ProtocolConfig, ) -> UserInputResult

fn check_sponsorship(&self) -> UserInputResult

Check if the transaction is compliant with sponsorship.

fn is_system_tx(&self) -> bool

fn is_genesis_tx(&self) -> bool

fn is_end_of_epoch_tx(&self) -> bool

returns true if the transaction is one that is specially sequenced to run at the very end of the epoch

fn is_sponsored_tx(&self) -> bool

Check if the transaction is sponsored (namely gas owner != sender)

fn sender_mut_for_testing(&mut self) -> &mut IotaAddress

fn gas_data_mut(&mut self) -> &mut GasData

fn expiration_mut_for_testing(&mut self) -> &mut TransactionExpiration